refactor(vote): use simple proposalVoting hook
