+++ plugins/maciVoting/hooks/useL2ProposalStage.tsx
@@ -3,7 +3,7 @@ import { ITransformedStage, IVote, ProposalStages } from "@/utils/types";
 import dayjs from "dayjs";
 import { formatEther } from "viem";
 import { useVotingToken } from "./useVotingToken";
-import { useProposalVoting } from "./useProposalVoting";
+import { useProposalVoting } from "./useProposalVotingOld";
 import { useProposalStatus } from "./useProposalVariantStatus";
 import { useGetProposalVotesL2 } from "./usePendingVotesRelay";
 import { useRelayVotesList } from "./useProposalVoteList";
